+++ NEWS.md
@@ -2,6 +2,8 @@
 * Added ker option to get_mrsi_voi function.
 * Added Asc, BHB, Cho, PEth and Ser simulation parameters.
 * Added append_basis function to combine two basis sets objects.
+* The align function now accepts more than one reference frequency and had a 
+change in the argment naming.
 
 # spant 0.17.0
 * Added a function to grid shift 2D MRSI data in the x/y direction.